Refactor packet structure and switch to uint16_t packet IDs
Replaced static packet ID definitions with `nsDefPacket` template for consistency. Transitioned packet ID types to `uint16_t` and updated related method signatures and data structures accordingly. Fixed minor logic issues and cleaned up deprecated code, improving maintainability and robustness.

+++ Networking/client/Client.h
@@ -28,8 +28,8 @@ public:
    void Connect(const char *ip, int port);
    void Disconnect();
 
-   void AddPacketHandler(int packetId, const nsPacketsHandlingManager::HandlerCallback& handler);
-   void ClearPacketHandler(int packetId);
+   void AddPacketHandler(uint16_t packetId, const nsPacketsHandlingManager::HandlerCallback& handler);
+   void ClearPacketHandler(uint16_t packetId);
    void SetCommonPacketsHandler(const PacketHandler& handler);
    void Update();
